+++ ppx/test/recursive_type_bindings.t
@@ -13,9 +13,9 @@
   >   | Base of 'a tree * 'a forest [@@deriving json]
   > EOF
     let rec a_of_json =
-    and b_of_json = (fun x -> int_of_json x : Yojson.Basic.t -> b)
+    and b_of_json = (fun x -> int_of_json x : Yojson.Safe.t -> b)
     let rec a_to_json =
-    and b_to_json = (fun x -> int_to_json x : b -> Yojson.Basic.t)
+    and b_to_json = (fun x -> int_to_json x : b -> Yojson.Safe.t)
     let rec of_json =
     let rec to_json =
     let rec lst_of_json a_of_json =